+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "reool"
-version = "0.6.1"
+version = "0.7.0"
 authors = ["Christian Douven <chridou@users.noreply.github.com>"]
 readme = "README.md"
 license = "Apache-2.0/MIT"
