+++ src/node_pool/mod.rs
@@ -2,14 +2,15 @@
 use std::time::Duration;
 
 use crate::connection_factory::ConnectionFactory;
-use crate::error::{InitializationError, InitializationResult};
+use crate::error::InitializationResult;
 use crate::executor_flavour::ExecutorFlavour;
 use crate::helpers;
 use crate::instrumentation::{Instrumentation, NoInstrumentation};
-use crate::pool::{Config as PoolConfig, Pool, Poolable};
-use crate::{Checkout, RedisPool};
+use crate::pool::{Config as PoolConfig, Pool};
+use crate::{Checkout, Poolable, RedisPool};
 
 pub use crate::backoff_strategy::BackoffStrategy;
+pub use crate::error::InitializationError;
 pub use crate::pool::PoolStats;
 
 /// A configuration for creating a `SingleNodePool`.
